+++ src/main/java/agh/ics/oop/model/EnergyOptions.java
@@ -1,6 +1,6 @@
 package agh.ics.oop.model;
 
-public record EnergyOptions(double energyFromPlant, double dailyEnergyLoss, double energyToReproduce, double energyToKid) {
+public record EnergyOptions(int energyFromPlant, int dailyEnergyLoss, int energyToReproduce, int energyToKid) {
     public EnergyOptions {
         if (energyToKid> energyToReproduce){
             throw new IllegalArgumentException("Energy that is obligatory to reproducing must be >= than energy that is given to kid");
